requirements.txt
Flask-RESTful
Flask-SQLAlchemy
psycopg2
alembic
Flask-Migrate
passlib
PyJWT
gunicorn

